Describe the changes made in Yolo
