+++ README.md
@@ -1,12 +1,17 @@
-# Yolo
-Documentation: 
+# Yolo Application Overview
 
-https://github.com/ultralytics/ultralytics?tab=readme-ov-file
+This application utilizes YOLO models for various monitoring and analysis purposes. It includes three main functionalities within the `models` folder:
 
-https://docs.ultralytics.com/
+- `detect_with_camera`: Provides real-time camera supervision.
+- `detect_with_display`: Monitors activities on your desktop.
+- `detect_with_files`: Analyzes pre-recorded videos.
 
-Install requirements: 
+## Installation
 
-pip install -r requirements.txt
+To set up the application and install the required dependencies, follow these steps:
 
-pip install ultralytics
+1. Install all required libraries from the requirements file:
+   ```bash
+   pip install -r requirements.txt
+2. install extra libraries:
+    pip install ultralytics
